chore: Rename orb client to anchor origin provider
Since we now have two orb clients(anchor origin provider and resolution verifier) we should move anchor origin provider into separate package orbclient/aoprovider. Also, move nsprovider and verprovider folders from orbclient to orbclient/protocol.

Closes #896
