+++ pkg/orbclient/protocol/nsprovider/namespaceprovider.go
